Comprehensive framework for IPv6 support
This is based on work I have done for the AstLinux project
(AstLinux.org).

IPv6 and IPv4 exist side-by-side:  The same network interface may have
both IPv4 and IPv6.  To support both, add the following options:

  * New `no` option for `use` to skip updating the IPv4 address.
  * `usev6`: One of: `no`, `ip`, `if`, `cmd`, `web`.
  * `ipv6`: The IPv6 address to use when `usev6=ip`.
  * `webv6`: The URL of a web server that will return the IPv6 address
    when `usev6=web`.
  * `webv6-skip`: The string to skip to before looking for the IPv6
    address in the `webv6` output.
  * `cmdv6`: The system command to execute that will return an IPv6
    address when `usev6=cmd`
  * `cmdv6-skip`: The string to skip to before looking for the IPv6
    address in the `cmdv6` output.

There is no `ifv6` option, as it is assumed that the same interface is
used for both IPv4 and IPv6.

You can mix methods for IPv4 and IPv6 for the same interface.  For
example, the following would be a valid method for a system behind an
IPv4 NAT connection:

    use=web
    usev6=if
    if=eth0
    web=http://ipdetect.dnspark.com/

With the above, the IPv6 address would be obtained from eth0 (globaly
routable address) and the public IPv4 address by contacting a web
site.

Also:
  * Add support for updating AAAA (IPv6) DNS records for
    freedns.afraid.org.
  * Update the existing IPv6 nsupdate method to account for the new
    design.
  * Add optional support is added for cURL as some embedded systems do
    not have Perl NET6 and SSL packages installed.  It is enabled with
    `curl=yes`.